+++ lib/accessibility/running_application.rb
@@ -1,0 +1,8 @@
+require 'accessibility/bridge'
+
+if on_macruby?
+  framework 'Cocoa'
+else
+  require 'accessibility/running_application.bundle'
+end
+