+++ History.markdown
@@ -1,3 +1,7 @@
+# v0.2.0 - Add some additional useful classes
+
+ * Add a 99% drop-in replacement for `NSRunningApplication`
+
 # v0.1.1 - Detect structs
 
  * Fix detection of struct type objects in `to_ax()`